update to version 0.14.44
This build also includes a completely refreshed set of BuildRequires and
Requires (since upstream likes to change what they depend on a lot).

The patches to the upstream build script were rebased (again). The plan is to
drop those patches and to inject upstream's custom build flags/tags directly
in the .spec file in the future, and to only use the upstream build script to
generate the resource files, not to build the binaries.

I had to disable the "NoSync" feature in the rewritten "stdiscosrv" database
backend, because the goleveldb package in fedora is too old to support it -
this probably leads to worse performance there.

I also re-enabled checks on some tests that were previously failing randomly.
If they re-surface at some point in the future, I'll ignore them again.